refactor: json array, object 타입에 따라 분기 처리

+++ real-estate-domain/src/main/java/io/gig/realestate/domain/realestate/construct/dto/ConstructDataApiDto.java
@@ -84,7 +84,7 @@ public class ConstructDataApiDto {
                 .bldNm(item.has("bldNm") ? item.getString("bldNm") : null)
                 .hhldCnt(item.has("hhldCnt") ? item.getInt("hhldCnt") : 0)
                 .useAprDay(item.has("useAprDay") ? item.getInt("useAprDay") : 0)
-                .platArea(item.has("platArea") ? item.getInt("platArea") : 0)
+                .platArea(item.has("platArea") ? item.getDouble("platArea") : 0)
                 .archArea(item.has("archArea") ? item.getDouble("archArea") : null)
                 .bcRat(item.has("bcRat") ? item.getDouble("bcRat") : null)
                 .totArea(item.has("totArea") ? item.getDouble("totArea") : null)
